+++ src/api/dm/addon/dsp/synthesis/effect_synthesis.d
@@ -10,15 +10,15 @@ import Math = api.math;
 struct ADSR
 {
     //sample * adsr(..)
-    double attack = 0.2;
-    double decay = 0.2;
-    double sustain = 0.7;
-    double release = 0.2;
+    float attack = 0.2;
+    float decay = 0.2;
+    float sustain = 0.7;
+    float release = 0.2;
 
-    double adsr(double time)
+    float adsr(float time)
     {
         //(Attack + Decay + Release) <= duration
-        const double releaseTime = 1 - release;
+        const float releaseTime = 1 - release;
 
         //Attack
         if (time < attack)
@@ -61,7 +61,7 @@ bool fadein(T)(T[] buffer, size_t samples, size_t channels = 2)
 
     for (auto i = 0; i < samples * channels; i += channels)
     {
-        double factor = ((cast(double) i) / channels) / samples;
+        float factor = ((cast(float) i) / channels) / samples;
 
         for (auto ch = 0; ch < channels; ch++)
         {
@@ -88,7 +88,7 @@ bool fadeout(T)(T[] buffer, size_t samples, size_t channels)
 
     for (auto i = start; i < buffer.length; i += channels)
     {
-        double factor = 1.0 - (((cast(double) i) - start) / channels) / samples;
+        float factor = 1.0 - (((cast(float) i) - start) / channels) / samples;
 
         for (size_t ch = 0; ch < channels; ch++)
         {
@@ -104,19 +104,19 @@ bool fadeout(T)(T[] buffer, size_t samples, size_t channels)
     return true;
 }
 
-double distortionHard(double sample, double threshold)
+float distortionHard(float sample, float threshold)
 {
     return Math.max(Math.min(sample, threshold), -threshold);
 }
 
-double distortionSoft(double sample, double gain = 0.3) => Math.tanh(sample * gain);
+float distortionSoft(float sample, float gain = 0.3) => Math.tanh(sample * gain);
 
-double distortionCube(double sample, double drive)
+float distortionCube(float sample, float drive)
 {
     return sample - (1.0 / 3.0) * sample * sample * sample * drive;
 }
 
-double lowpass(double prevSample, double currentSample, double cutoff)
+float lowpass(float prevSample, float currentSample, float cutoff)
 {
     return prevSample + cutoff * (currentSample - prevSample); // cutoff ~0.1-0.3
 }
@@ -124,17 +124,17 @@ double lowpass(double prevSample, double currentSample, double cutoff)
 //Low-Frequency Oscillator
 //tremolo: output = signal * lfoValue
 //vibrato: freq = base_freq + lfo_value * depth)
-double lfo(double freq, double time)
+float lfo(float freq, float time)
 {
     return 0.5 * (1.0 + Math.sin(Math.PI2 * freq * time));
 }
 
-double tremolo(double sample, double sampleTime, double freq = 5)
+float tremolo(float sample, float sampleTime, float freq = 5)
 {
     return sample * lfo(freq, sampleTime);
 }
 
-double vibrato(double sampleTime, double baseFreq, double depthFromRate = 0.01, double rateHz = 6.0)
+float vibrato(float sampleTime, float baseFreq, float depthFromRate = 0.01, float rateHz = 6.0)
 {
     auto freq = baseFreq * (1.0 + depthFromRate * Math.sin(Math.PI2 * rateHz * sampleTime));
     return Math.sin(Math.PI2 * freq * sampleTime);
@@ -147,7 +147,7 @@ double vibrato(double sampleTime, double baseFreq, double depthFromRate = 0.01, 
  * up-down, C → G → E → C → G...	
  * random, E → C → G → E...
  */
-double arpeggio(double time, double sample)
+float arpeggio(float time, float sample)
 {
     if (time < 1.0)
         sample += Math.sin(2 * Math.PI * 261.63 * time) * 0.3; // C4
