+++ src/api/dm/addon/dsp/synthesis/signal_synthesis.d
@@ -7,22 +7,22 @@ import Math = api.math;
 /**
  * Authors: initkfs
  */
-double sine(double time, double freq, double phase)
+float sine(float time, float freq, float phase)
 {
     //phase == [0, 1)
     return Math.sin(Math.PI2 * freq * time + phase);
 }
 
-double sinovertones(double time, double freq, double phase)
+float sinovertones(float time, float freq, float phase)
 {
-    double sample = Math.sin(Math.PI2 * freq * time + phase) * 0.7;
+    float sample = Math.sin(Math.PI2 * freq * time + phase) * 0.7;
     sample += Math.sin(Math.PI2 * 2.0 * freq * time + 2 * phase) * 0.3;
     sample += Math.sin(Math.PI2 * 3.0 * freq * time + 3 * phase) * 0.1;
     sample += Math.sin(Math.PI2 * 4.0 * freq * time + 4 * phase) * 0.05;
     return sample;
 }
 
-double square(double time, double freq, double phase)
+float square(float time, float freq, float phase)
 {
     import std.math : fmod;
 
@@ -30,14 +30,14 @@ double square(double time, double freq, double phase)
     //return (fmod(freq * time + phase, 1.0) < 0.5) ? 1.0 : -1.0;
 }
 
-double sawtooth(double time, double freq, double phase)
+float sawtooth(float time, float freq, float phase)
 {
     import std.math : fmod;
 
     return fmod(time * freq + phase, 1.0);
 }
 
-double triangle(double time, double freq, double phase)
+float triangle(float time, float freq, float phase)
 {
     //return 1.0 - 4.0 * Math.abs(fmod(freq * time + phase, 1.0) - 0.5);
     return (2.0 * Math.abs(2.0 * ((time * freq + phase) - Math.floor(time * freq + 0.5))) - 1.0);
@@ -45,48 +45,48 @@ double triangle(double time, double freq, double phase)
 
 struct FMdata
 {
-    double fc = 0;
-    double fm = 0;
-    double index = 0;
-    double durationMs = 0;
+    float fc = 0;
+    float fm = 0;
+    float index = 0;
+    float durationMs = 0;
     bool isFcMulFm;
 }
 
-double fmodulator(double time, double phase, double fc, double fm, double index)
+float fmodulator(float time, float phase, float fc, float fm, float index)
 {
     return Math.sin((Math.PI2 * fc * time + phase) + index * Math.sin(Math.PI2 * fm * time));
 }
 
 import api.math.random : Random;
 
-double whiteNoise(Random rnd)
+float whiteNoise(Random rnd)
 {
-    return rnd.between!double(-1, 1);
+    return rnd.between!float(-1, 1);
 }
 
-// void chaoticLogistic(T)(T[] buffer, double r, double sampleRateHz, double amplitude0to1 = 1.0, size_t channels = 2)
+// void chaoticLogistic(T)(T[] buffer, float r, float sampleRateHz, float amplitude0to1 = 1.0, size_t channels = 2)
 // {
-//     double x = 0.5;
+//     float x = 0.5;
 //     onBuffer(buffer, sampleRateHz, amplitude0to1, channels, (i, time) {
 //         x = r * x * (1.0 - x);
 //         return (x - 0.5f) * 2.0;
 //     });
 // }
 
-// void poissonNoise(T)(Random rnd, T[] buffer, double rate, double sampleRateHz, double amplitude0to1 = 1.0, size_t channels = 2)
+// void poissonNoise(T)(Random rnd, T[] buffer, float rate, float sampleRateHz, float amplitude0to1 = 1.0, size_t channels = 2)
 // {
-//     double threshold = 1.0 - rate / sampleRateHz;
+//     float threshold = 1.0 - rate / sampleRateHz;
 
 //     onBuffer(buffer, sampleRateHz, amplitude0to1, channels, (i, time) {
 //         return (rnd.between0to1 > threshold) ? 1 : 0;
 //     });
 // }
 
-// void fractalNoise(T)(Random rnd, T[] buffer, double sampleRateHz, double amplitude0to1 = 1.0, size_t channels = 2)
+// void fractalNoise(T)(Random rnd, T[] buffer, float sampleRateHz, float amplitude0to1 = 1.0, size_t channels = 2)
 // {
 //     import core.stdc.stdlib : malloc, free;
 
-//     void* tempPtr = malloc(buffer.length * double.sizeof);
+//     void* tempPtr = malloc(buffer.length * float.sizeof);
 //     assert(tempPtr);
 
 //     scope (exit)
@@ -94,10 +94,10 @@ double whiteNoise(Random rnd)
 //         free(tempPtr);
 //     }
 
-//     double[] temp = (cast(double*) tempPtr)[0 .. buffer.length];
+//     float[] temp = (cast(float*) tempPtr)[0 .. buffer.length];
 //     assert(buffer.length == temp.length);
 
-//     double value = 0;
+//     float value = 0;
 
 //     foreach (ref t; temp)
 //     {
@@ -106,7 +106,7 @@ double whiteNoise(Random rnd)
 //         t = value;
 //     }
 
-//     double max = 0.0;
+//     float max = 0.0;
 //     foreach (ref t; temp)
 //     {
 //         const abst = Math.abs(t);
@@ -121,48 +121,48 @@ double whiteNoise(Random rnd)
 //     });
 // }
 
-// void phaseDistortion(T)(T[] buffer, double freqHz, double distortion, double sampleRateHz, double amplitude0to1 = 1.0, size_t channels = 2)
+// void phaseDistortion(T)(T[] buffer, float freqHz, float distortion, float sampleRateHz, float amplitude0to1 = 1.0, size_t channels = 2)
 // {
 //     import std.math : fmod;
 
-//     double phaseDt = freqHz / sampleRateHz;
-//     double phase = 0;
+//     float phaseDt = freqHz / sampleRateHz;
+//     float phase = 0;
 
 //     onBuffer(buffer, sampleRateHz, amplitude0to1, channels, (i, time) {
 //         phase = fmod(phase + phaseDt, 1.0);
-//         double distortPhase = phase < distortion ? phase / distortion : (
+//         float distortPhase = phase < distortion ? phase / distortion : (
 //             1.0 - phase) / (1.0 - distortion);
 //         return Math.sin(2.0f * Math.PI * distortPhase);
 //     });
 // }
 
-// void polynomialWaveshaper(T)(T[] buffer, double freqHz, double a, double b, double c, double sampleRateHz, double amplitude0to1 = 1.0, size_t channels = 2)
+// void polynomialWaveshaper(T)(T[] buffer, float freqHz, float a, float b, float c, float sampleRateHz, float amplitude0to1 = 1.0, size_t channels = 2)
 // {
-//     double phaseDt = freqHz / sampleRateHz;
-//     double phase = 0;
+//     float phaseDt = freqHz / sampleRateHz;
+//     float phase = 0;
 
 //     onBuffer(buffer, sampleRateHz, amplitude0to1, channels, (i, time) {
 //         phase = fmod(phase + phaseDt, 1.0);
 //         dobule x = 2.0 * phase - 1.0;
-//         double y = a * x + b * x * x * x + c * x * x * x * x * x;
+//         float y = a * x + b * x * x * x + c * x * x * x * x * x;
 //         return y;
 //     });
 // }
 
-// void diracImpulses(T)(T[] buffer, size_t period, double sampleRateHz, double amplitude0to1 = 1.0, size_t channels = 2)
+// void diracImpulses(T)(T[] buffer, size_t period, float sampleRateHz, float amplitude0to1 = 1.0, size_t channels = 2)
 // {
 //     onBuffer(buffer, sampleRateHz, amplitude0to1, channels, (i, time) {
 //         return (i % period == 0) ? 1 : 0;
 //     });
 // }
 
-// void am(T)(T[] buffer, double fc, double fm, double sampleRateHz, double amplitude0to1 = 1.0, size_t channels = 2)
+// void am(T)(T[] buffer, float fc, float fm, float sampleRateHz, float amplitude0to1 = 1.0, size_t channels = 2)
 // {
-//     double carrierDt = fc / sampleRateHz;
-//     double modDt = fm / sampleRateHz;
+//     float carrierDt = fc / sampleRateHz;
+//     float modDt = fm / sampleRateHz;
 
-//     double carrierPhase = 0;
-//     double modPhase = 0;
+//     float carrierPhase = 0;
+//     float modPhase = 0;
 
 //     import std.math : fmod;
 
@@ -171,8 +171,8 @@ double whiteNoise(Random rnd)
 //         carrierPhase = fmod(carrierPhase + carrierDt, 1.0);
 //         modPhase = fmod(modPhase + modDt, 1.0);
 
-//         double mod = 0.5 * (1.0 + Math.sin(Math.PI2 * modPhase));
-//         double carrier = Math.sin(Math.PI2 * carrierPhase);
+//         float mod = 0.5 * (1.0 + Math.sin(Math.PI2 * modPhase));
+//         float carrier = Math.sin(Math.PI2 * carrierPhase);
 
 //         return mod * carrier;
 //     });
